@@ -101,21 +101,7 @@ class Paystack
         );
     }
 
-    /**
-     * Get the authorization url from the callback response
-     * @return Paystack
-     * @throws IsNullException
-     */
-    public function getAuthorizationUrl()
-    {
-        $this->makePaymentRequest();
-
-        $this->url = $this->getResponse()['data']['authorization_url'];
-
-        return $this;
-    }
-
-    /**
+     /**
      * Initiate a payment request to Paystack
      * Included the option to pass the payload to this method for situations
      * when the payload is built on the fly (not passed to the controller from a view)
@@ -125,27 +111,39 @@ class Paystack
      */
     public function makePaymentRequest($data = null)
     {
-        if ($data == null) {
+        if ( $data == null ) {
+
+            $quantity = intval(request()->quantity);
+
             $data = [
-                "amount" => intval(request()->amount),
+                "amount" => intval(request()->amount) * $quantity,
                 "reference" => request()->reference,
                 "email" => request()->email,
                 "plan" => request()->plan,
                 "first_name" => request()->first_name,
                 "last_name" => request()->last_name,
                 "callback_url" => request()->callback_url,
+                "currency" => (request()->currency != ""  ? request()->currency : "NGN"),
                 /*
-                 * to allow use of metadata on Paystack dashboard and a means to return additional data back to redirect url
-                 * form need an input field:
-                 * <input type="hidden" name="metadata" value="{{ json_encode($array) }}"/>
-                 * array must be set up as:
-                 * $array = [ 'custom_fields' => [
-                 *               ['display_name' => "Cart Id", "variable_name" => "cart_id", "value" => "2"],
-                 *               ['display_name' => "Sex", "variable_name" => "sex", "value" => "female"]
-                 * ...
-                 *      ]
-                 *  ]
-                 */
+                    Paystack allows for transactions to be split into a subaccount -
+                    The following lines trap the subaccount ID - as well as the ammount to charge the subaccount (if overriden in the form)
+                    both values need to be entered within hidden input fields
+                */
+                "subaccount" => request()->subaccount,
+                "transaction_charge" => request()->transaction_charge,
+                /*
+                * to allow use of metadata on Paystack dashboard and a means to return additional data back to redirect url
+                * form need an input field: <input type="hidden" name="metadata" value="{{ json_encode($array) }}" >
+                *array must be set up as: $array = [ 'custom_fields' => [
+                *                                                            ['display_name' => "Cart Id", "variable_name" => "cart_id", "value" => "2"],
+                *                                                            ['display_name' => "Sex", "variable_name" => "sex", "value" => "female"],
+                *                                                            .
+                *                                                            .
+                *                                                            .
+                *                                                        ]
+                *
+                *                                  ]
+                */
                 'metadata' => request()->metadata
             ];
 
@@ -188,7 +186,7 @@ class Paystack
         return json_decode($this->response->getBody(), true);
     }
 
-    /**
+     /**
      * Get the authorization callback response
      * In situations where Laravel serves as an backend for a detached UI, the api cannot redirect
      * and might need to take different actions based on the success or not of the transaction
@@ -358,7 +356,7 @@ class Paystack
 
         $this->setRequestOptions();
 
-        $this->setHttpResponse("/plan", 'POST', $data);
+        return $this->setHttpResponse("/plan", 'POST', $data)->getResponse();
 
     }
 
@@ -476,7 +474,7 @@ class Paystack
         ];
 
         $this->setRequestOptions();
-        $this->setHttpResponse('/subscription', 'POST', $data);
+        return $this->setHttpResponse('/subscription', 'POST', $data)->getResponse();
     }
 
     /**
@@ -576,7 +574,7 @@ class Paystack
         ];
 
         $this->setRequestOptions();
-        $this->setHttpResponse('/page', 'POST', $data);
+        return $this->setHttpResponse('/page', 'POST', $data)->getResponse();
     }
 
     /**
@@ -627,8 +625,7 @@ class Paystack
      * @throws IsNullException
      */
 
-    public function createSubAccount()
-    {
+    public function createSubAccount(){
         $data = [
             "business_name" => request()->business_name,
             "settlement_bank" => request()->settlement_bank,
@@ -677,11 +674,11 @@ class Paystack
 
     /**
      * Updates a subaccount to be used for split payments . Required params are business_name , settlement_bank , account_number , percentage_charge
-     * @param string subaccount_code
+     * @param $subAccountCode
      * @return array
      * @throws IsNullException
      */
-    public function updateSubAccount($subaccount_code)
+    public function updateSubAccount($subAccountCode)
     {
         $data = [
             "business_name" => request()->business_name,
@@ -697,7 +694,7 @@ class Paystack
         ];
 
         $this->setRequestOptions();
-        return $this->setHttpResponse("/subaccount/{$subaccount_code}", "PUT", array_filter($data))->getResponse();
+        return $this->setHttpResponse("/subaccount/{$subAccountCode}", "PUT", array_filter($data))->getResponse();
 
     }
 }
